passing tempalte editor features with tempalte editor
