+++ src/calls/embedTemplateEditorPage.ts
@@ -4,6 +4,11 @@ import {createEmbedIframe} from "../utils/iframe.ts";
 import {validatePermissions} from "../helpers/permissions.ts";
 import {parseTokenPayload} from "../helpers/token.ts";
 import {type SdkPermission} from "../helpers/permissions.ts";
+import {
+  type TemplateEditorFeatures,
+  TEMPLATE_EDITOR_FEATURES_NONE,
+  TEMPLATE_EDITOR_FEATURES_DEFAULT,
+} from "../helpers/embedFeatures.ts";
 
 export interface EmbedTemplateEditorPageOptions {
   /**
@@ -26,28 +31,24 @@ export function embedTemplateEditorPage(
   const {workspaceHandle, permissions} = parseTokenPayload(sdk.token);
 
   validatePermissions(permissions, REQUIRED_PERMISSIONS);
-  validateFeatures(features ?? {});
 
   const iframe = createEmbedIframe({
     token: sdk.token,
     path: sdk.path,
     pathHash: `/${workspaceHandle}/templates/${templateId}/edit`,
-    config: {features, fonts, appearance},
+    config: {
+      features: {
+        templateEditor:
+          Object.keys(features ?? {}).length > 0
+            ? {...TEMPLATE_EDITOR_FEATURES_NONE, ...features}
+            : TEMPLATE_EDITOR_FEATURES_DEFAULT,
+      },
+      fonts,
+      appearance,
+    },
   });
 
   element.replaceChildren(iframe);
-}
-
-export interface TemplateEditorConfig {
-  features?: TemplateEditorFeatures | undefined;
-  fonts?: FontSource[] | undefined;
-  appearance?: AppearanceConfig | undefined;
-}
-
-function validateFeatures(features: TemplateEditorFeatures): void {
-  if (features && Object.keys(features).length === 0) {
-    throw new Error("please specify at least one template editor feature.");
-  }
 }
 
 const REQUIRED_PERMISSIONS: SdkPermission[][] = [
@@ -55,89 +56,3 @@ const REQUIRED_PERMISSIONS: SdkPermission[][] = [
   ["user:read", "user:write"],
   ["template:write", "template:read"],
 ];
-
-export interface TemplateEditorFeatures {
-  logo?: boolean;
-  icon?: boolean;
-  heading?: boolean;
-  colors?: boolean;
-  coverImage?: boolean;
-  headerField?: boolean;
-  secondaryFields?: boolean;
-  barcodeType?: boolean;
-  barcodeData?: boolean;
-  barcodeText?: boolean;
-  description?: boolean;
-  backFields?: boolean;
-  locationMessages?: boolean;
-  appLinks?: boolean;
-  expiry?: boolean;
-}
-
-export const TEMPLATE_EDITOR_FEATURES_DEFAULT = {
-  logo: true,
-  icon: true,
-  heading: true,
-  colors: true,
-  coverImage: true,
-  headerField: true,
-  secondaryFields: true,
-  barcodeType: false,
-  barcodeData: false,
-  barcodeText: false,
-  description: false,
-  backFields: true,
-  locationMessages: true,
-  appLinks: true,
-  expiry: false,
-} satisfies Record<keyof TemplateEditorFeatures, true | false>;
-
-export const TEMPLATE_EDITOR_FEATURES_ALL: Record<
-  keyof TemplateEditorFeatures,
-  true
-> = {
-  logo: true,
-  icon: true,
-  heading: true,
-  colors: true,
-  coverImage: true,
-  headerField: true,
-  secondaryFields: true,
-  barcodeType: true,
-  barcodeData: true,
-  barcodeText: true,
-  description: true,
-  backFields: true,
-  locationMessages: true,
-  appLinks: true,
-  expiry: true,
-};
-
-export const TEMPLATE_EDITOR_FEATURES_NONE: Record<
-  keyof TemplateEditorFeatures,
-  false
-> = {
-  logo: false,
-  icon: false,
-  heading: false,
-  colors: false,
-  coverImage: false,
-  headerField: false,
-  secondaryFields: false,
-  barcodeType: false,
-  barcodeData: false,
-  barcodeText: false,
-  description: false,
-  backFields: false,
-  locationMessages: false,
-  appLinks: false,
-  expiry: false,
-};
-
-export const TEMPLATE_EDITOR_FEATURES_ENABLE_BARCODE: Partial<
-  Record<keyof TemplateEditorFeatures, true>
-> = {
-  barcodeType: true,
-  barcodeData: true,
-  barcodeText: true,
-};
